+++ VadereGui/src/org/vadere/gui/topographycreator/control/ActionMergeObstacles.java
@@ -1,11 +1,13 @@
 package org.vadere.gui.topographycreator.control;
 
+import org.apache.commons.lang3.tuple.Pair;
 import org.vadere.gui.topographycreator.model.IDrawPanelModel;
+import org.vadere.simulator.projects.migration.GeometryCleaner;
 import org.vadere.state.attributes.scenario.AttributesObstacle;
 import org.vadere.state.scenario.Obstacle;
-import org.vadere.meshing.WeilerAtherton;
 import org.vadere.util.geometry.shapes.VPolygon;
 import org.vadere.util.geometry.shapes.VRectangle;
+import org.vadere.util.logging.Logger;
 
 import java.awt.event.ActionEvent;
 import java.util.ArrayList;
@@ -22,6 +24,7 @@ import javax.swing.undo.UndoableEditSupport;
 public class ActionMergeObstacles extends TopographyAction {
 
 	private final UndoableEditSupport undoSupport;
+	private static Logger logger = Logger.getLogger(ActionMergeObstacles.class);
 
 	public ActionMergeObstacles(String name, ImageIcon icon, IDrawPanelModel panelModel,
 	                            UndoableEditSupport undoSupport) {
@@ -41,19 +44,35 @@ public class ActionMergeObstacles extends TopographyAction {
 				.map(shape -> ((VPolygon)shape))
 				.collect(Collectors.toList());
 
-		WeilerAtherton weilerAtherton = new WeilerAtherton(polygons);
-		List<VPolygon> mergedPolygons = weilerAtherton.cup();
+		GeometryCleaner topographyCleaner = new GeometryCleaner(
+				new VRectangle(getScenarioPanelModel().getTopographyBound()), polygons, 0.01);
+
+		logger.info("start merging, this can require some time.");
+		Pair<VPolygon, List<VPolygon>> mergedPolygons = topographyCleaner.clean();
+		//polygons = WeilerAtherton.magnet(polygons, new VRectangle(getScenarioPanelModel().getBounds()));
+		logger.info("merging process finisehd.");
+
+
+
+		/*WeilerAtherton weilerAtherton = new WeilerAtherton(polygons);
+		List<VPolygon> mergedPolygons = weilerAtherton.cup();*/
 
 		// remove polygon obstacles
 		getScenarioPanelModel().removeObstacleIf(obstacle ->
 				obstacle.getShape() instanceof VPolygon || obstacle.getShape() instanceof VRectangle);
 
+
+
 		// add merged obstacles
-		mergedPolygons
+		mergedPolygons.getRight()
 				.stream()
 				.map(polygon -> new Obstacle(new AttributesObstacle(-1, polygon)))
 				.forEach(obstacle -> getScenarioPanelModel().addShape(obstacle));
 
+		//WeilerAtherton weilerAtherton = new WeilerAtherton(Arrays.asList(mergedPolygons.getLeft(), new VPolygon(getScenarioPanelModel().getBounds())));
+		//weilerAtherton.cap().get();
+
+		//getScenarioPanelModel().addShape(new Obstacle(new AttributesObstacle(-1, weilerAtherton.cap().get())));
 
 		List<Obstacle> after = new ArrayList<>(getScenarioPanelModel().getObstacles());