+++ VadereSimulator/src/org/vadere/simulator/entrypoints/cmd/commands/UtilsSubCommand.java
@@ -5,6 +5,7 @@ import net.sourceforge.argparse4j.inf.Namespace;
 
 import org.apache.commons.lang3.tuple.Pair;
 import org.vadere.simulator.entrypoints.ScenarioFactory;
+import org.vadere.util.version.Version;
 import org.vadere.simulator.entrypoints.cmd.SubCommandRunner;
 import org.vadere.simulator.models.potential.fields.IPotentialField;
 import org.vadere.simulator.models.potential.fields.PotentialFieldDistancesBruteForce;
@@ -37,7 +38,6 @@ public class UtilsSubCommand implements SubCommandRunner {
 		methods.put("getHash", Pair.of("[-i: file, -o: ignored]", this::getHash));
 		methods.put("binCache", Pair.of("[-i: file, -o: directory]",this::calculateBinCache));
 		methods.put("txtCache", Pair.of("[-i: file, -o: directory]",this::calculateTextCache));
-
 	}
 
 	public String[] methodsString(){
@@ -85,10 +85,8 @@ public class UtilsSubCommand implements SubCommandRunner {
 	 *  Return the hash used to check if a new cache value needed to be calculated.
 	 */
 	private void getHash(Namespace ns, ArgumentParser parser) throws Exception{
-		Scenario scenario = createScenario(ns.getString("input"));
-		System.out.println(ScenarioCache.getHash(scenario));
+		calculateCache(ns, parser, CacheType.TXT_CACHE);
 	}
-
 
 	private void calculateBinCache(Namespace ns, ArgumentParser parser) throws Exception{
 		calculateCache(ns, parser, CacheType.BIN_CACHE);