+++ VadereSimulator/src/org/vadere/simulator/projects/migration/incident/IncidentMigrationAssistant.java
@@ -2,7 +2,7 @@ package org.vadere.simulator.projects.migration.incident;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
-import org.vadere.simulator.entrypoints.Version;
+import org.vadere.util.version.Version;
 import org.vadere.simulator.projects.migration.MigrationAssistant;
 import org.vadere.simulator.projects.migration.MigrationException;
 import org.vadere.simulator.projects.migration.MigrationOptions;