+++ VadereSimulator/tests/org/vadere/simulator/projects/migration/jsontranformation/JoltTransformV5toV6Test.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 
 import org.junit.Test;
-import org.vadere.simulator.entrypoints.Version;
+import org.vadere.util.version.Version;
 import org.vadere.simulator.projects.migration.MigrationException;
 
 import java.nio.file.Path;